+++ src/components/Header.scss
@@ -1,0 +1,86 @@
+@import "../styles/shared.scss";
+
+/* Menu */
+.portal-menu {
+  text-align: center;
+  z-index: 9; /* TODO constant for this? */
+  width: $menu-width;
+  display: flex;
+}
+.portal-menu-content {
+  display: flex;
+}
+
+.portal-menu-item {
+  margin: auto;
+  font-family: RajdhaniBold;
+  text-align: center;
+}
+
+.portal-title {
+  line-height: 72px;
+  font-size: 72px;
+  text-align: center;
+  font-family: MontserratBold;
+  color: white;
+  padding: 20px;
+}
+
+.portal-header {
+  width: 100%;
+  display: flex;
+  background-color: $menu-bar;
+}
+
+// title colors
+.title-P {
+ color: $activities-color; 
+}
+
+.title-O {
+  color: $materials-color; 
+ }
+
+ .title-R {
+  color: $help-color; 
+ }
+
+ .title-T {
+  color: $about-color; 
+ }
+
+ .title-A {
+  color: $profile-color; 
+ }
+
+ .title-L {
+  color: $button-color; 
+ }
+
+ //pages
+ .portal-menu-item-activities {
+   color: $activities-color;
+ }
+
+ .portal-menu-item-materials {
+  color: $materials-color;
+}
+
+.portal-menu-item-help {
+  color: $help-color;
+}
+
+.portal-menu-item-about {
+  color: $about-color;
+}
+
+.portal-menu-item-profile {
+  color: $profile-color;
+}
+
+.portal-menu-icon {
+  height: 60px;
+  width: 60px;
+  border-radius: 50%;
+  margin: 5px;
+}
