--- src/components/Menu.scss
@@ -1,95 +1,0 @@
-/* Menu */
-
-@import '../styles/shared.scss';
-
-@font-face {
-  font-family: menuFont;
-  src: url(../fonts/Chewy.ttf);
-}
-
-@font-face {
-  font-family: MontserratRegular;
-  src: url(../fonts/Montserrat/Montserrat-Regular.ttf);
-}
-
-@font-face {
-  font-family: MontserratBold;
-  src: url(../fonts/Montserrat/Montserrat-Bold.ttf);
-}
-
-@font-face {
-  font-family: RajdhaniRegular;
-  src: url(../fonts/Rajdhani/Rajdhani-Regular.ttf);
-}
-
-@font-face {
-  font-family: RajdhaniBold;
-  src: url(../fonts/Rajdhani/Rajdhani-Bold.ttf);
-}
-
-.portal-menu {
-  // position: absolute;
-  // top: 0;
-  // left: 5%;
-  text-align: center;
-  z-index: 9; /* TODO constant for this? */
-  width: $menu-width;
-  display: flex;
-  // flex-direction: column;
-}
-
-.menu {
-  width: 100%;
-  display: flex;
-  margin: auto;
-}
-
-.portal-menu-content {
-  display: flex;
-}
-
-.portal-menu-item {
-  margin: auto;
-}
-
-// .portal-menu-title {
-//   font-size: 24px;
-//   cursor: pointer;
-//   margin: 0 auto;
-// }
-
-// .portal-menu-content {
-//   background-color: white;
-//   // border: 3px outset white;
-//   border: 1px solid $primary-brown;
-//   box-shadow: 1px 1px 2px;
-// }
-
-// .portal-menu-item {
-//   // height: 50px;
-//   font-size: 13.5pt;
-//   margin: 5px 15px 0px 15px;
-//   padding: 6px 6px;
-//   border-bottom: 2px solid gray;
-//   font-family: menuFont;
-// }
-
-// .portal-menu-item-home {
-//   color: $menu-home;
-// }
-
-// .portal-menu-item-resources {
-//   color: $menu-resources;
-// }
-
-// .portal-menu-item-activities {
-//   color: $menu-activities;
-// }
-
-// .portal-menu-item-access  {
-//   color: $menu-access;
-// }
-
-// .portal-menu-item-profile  {
-//   color: $menu-profile;
-// }
