+++ src/styles/shared.scss
@@ -17,3 +17,39 @@ $orange: #ffbd4a;
 // Stories
 $stories-text-background: green;
 $stories-text: white;
+
+// new defined colors
+$activities-color: #79b4b3;
+$materials-color: #a9bb59;
+$help-color: #fa8e47;
+$about-color: #fc5e5a;
+$profile-color: #a586c5;
+$button-color: #f5bf53;
+$menu-bar: #201d1a;
+
+
+// fonts
+@font-face {
+  font-family: menuFont;
+  src: url(/fonts/Chewy.ttf);
+}
+
+@font-face {
+  font-family: MontserratRegular;
+  src: url(/fonts/Montserrat/Montserrat-Regular.ttf);
+}
+
+@font-face {
+  font-family: MontserratBold;
+  src: url(/fonts/Montserrat/Montserrat-Bold.ttf);
+}
+
+@font-face {
+  font-family: RajdhaniRegular;
+  src: url(/fonts/Rajdhani/Rajdhani-Regular.ttf);
+}
+
+@font-face {
+  font-family: RajdhaniBold;
+  src: url(/fonts/Rajdhani/Rajdhani-Bold.ttf);
+}
